# business/tecma-digital-platform/followup-3.0/be-followup-v3/openapi/openapi.v1.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Followup 3.0 REST API",
    "version": "1.0.0",
    "description": "API REST per Followup 3.0 (CRM multi-progetto). Gestione clienti, appartamenti, trattative (rent/sell), calendario, workspaces, progetti, audit e report. Autenticazione JWT (login, refresh). In locale il backend è su porta 8080."
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1",
      "description": "Backend locale (dev)"
    },
    {
      "url": "/v1",
      "description": "Stesso host (proxy)"
    }
  ],
...
  "paths": {
...
    "/workspaces/{workspaceId}/assets/upload-url": {
      "post": {
        "tags": [
          "assets",
          "workspaces"
        ],
        "summary": "URL presigned upload asset",
        "operationId": "postWorkspaceAssetUploadUrl",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "ws_demo_001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Metadati file e tipo asset"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URL e campi presigned"
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato"
          },
          "500": {
            "description": "Errore server"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/assets": {
      "get": {
        "tags": [
          "assets",
          "workspaces"
        ],
        "summary": "Lista asset workspace",
        "operationId": "listWorkspaceAssets",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "ws_demo_001"
          },
          {
            "name": "projectId",
            "in": "query",
            "required": false,
            "description": "Filtro progetto",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apartmentId",
            "in": "query",
            "required": false,
            "description": "Filtro appartamento",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "description": "Tipo asset",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista in data"
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato"
          },
          "500": {
            "description": "Errore server"
          }
        }
      },
      "post": {
        "tags": [
          "assets",
          "workspaces"
        ],
        "summary": "Registra asset dopo upload",
        "operationId": "createWorkspaceAsset",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "ws_demo_001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Metadati asset creato"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Asset creato in data"
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato"
          },
          "500": {
            "description": "Errore server"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/assets/{assetId}/download-url": {
      "get": {
        "tags": [
          "assets",
          "workspaces"
        ],
        "summary": "URL download presigned asset",
        "operationId": "getWorkspaceAssetDownloadUrl",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "ws_demo_001"
          },
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "description": "ID asset",
            "schema": {
              "type": "string"
            },
            "example": "673abc12def4567890123456"
          }
        ],
        "responses": {
          "200": {
            "description": "URL download"
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato"
          },
          "500": {
            "description": "Errore server"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/assets/{assetId}": {
      "delete": {
        "tags": [
          "assets",
          "workspaces"
        ],
        "summary": "Elimina asset",
        "operationId": "deleteWorkspaceAsset",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "ws_demo_001"
          },
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "description": "ID asset",
            "schema": {
              "type": "string"
            },
            "example": "673abc12def4567890123456"
          }
        ],
        "responses": {
          "200": {
            "description": "Eliminato"
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato"
          },
          "500": {
            "description": "Errore server"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from login or refresh"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Platform consumer API key (scoped by workspace/project)."
      },
      "signatureWebhookAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-signature-webhook-secret",
        "description": "Segreto condiviso per webhook firme (stesso valore di SIGNATURE_WEBHOOK_SECRET). Equivalente: Authorization Bearer <secret>."
      }
    },
...
  }
}